+++ cmd/root.go
@@ -10,7 +10,8 @@ import (
 	"github.com/spf13/cobra"
 
 	"github.com/giantswarm/backstage-catalog-importer/cmd/appcatalogs"
-	"github.com/giantswarm/backstage-catalog-importer/cmd/users"
+	installations "github.com/giantswarm/backstage-catalog-importer/cmd/installations"
+	users "github.com/giantswarm/backstage-catalog-importer/cmd/users"
 	"github.com/giantswarm/backstage-catalog-importer/pkg/input/helmchart"
 	"github.com/giantswarm/backstage-catalog-importer/pkg/input/repositories"
 	"github.com/giantswarm/backstage-catalog-importer/pkg/input/teams"
@@ -40,6 +41,7 @@ func init() {
 	rootCmd.PersistentFlags().StringP("output", "o", ".", "Output directory path")
 
 	rootCmd.AddCommand(appcatalogs.Command)
+	rootCmd.AddCommand(installations.Command)
 	rootCmd.AddCommand(users.Command)
 }