+++ pkg/output/bscatalog/v1alpha1/resource.go
@@ -1,0 +1,8 @@
+package v1alpha1
+
+type ResourceSpec struct {
+	Owner     string   `yaml:"owner"`
+	Type      string   `yaml:"type"`
+	System    string   `yaml:"system,omitempty"`
+	DependsOn []string `yaml:"dependsOn,omitempty"`
+}