+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add `installations` command to export entities for Giant Swarm installations.
+
 ## [0.14.0] - 2024-07-10
 
 ### Added